Tried to get Maven out of the Show View menu.
Ran into a bug with org.eclipse.e4.ui.workbench.swt that requires
org.eclipse.equinox.ds to be added manually to the plug-in. This causes
org.eclipse.m2e dependencies to be added.  These are where the Maven is
coming from.  Got it working and deployed but with Maven views.